+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   New login page (#220)
 -   Ability to pin compute plans to the top of the table (#222)
+-   About modal that opens from the user menu (#223)
 
 ### Fixed